Rename burrow_evm_client package to client
https://blog.golang.org/package-names
