+++ families/burrow_evm/src/sawtooth_burrow_evm/cli/exec.go
@@ -20,7 +20,7 @@ package main
 import (
 	"fmt"
 	"github.com/jessevdk/go-flags"
-	client "sawtooth_burrow_evm/client"
+	"sawtooth_burrow_evm/client"
 )
 
 type Exec struct {
