+++ families/burrow_evm/src/sawtooth_burrow_evm/cli/show.go
@@ -20,7 +20,7 @@ package main
 import (
 	"fmt"
 	"github.com/jessevdk/go-flags"
-	client "sawtooth_burrow_evm/client"
+	"sawtooth_burrow_evm/client"
 	. "sawtooth_burrow_evm/protobuf/evm_pb2"
 	sdk "sawtooth_sdk/client"
 )
